Order sync delete articles not in order

+++ database/migrations/2022_07_31_135200_add_skryfall_data_to_cards.php
@@ -14,7 +14,7 @@ class AddSkryfallDataToCards extends Migration
     public function up()
     {
         Schema::table('cards', function (Blueprint $table) {
-            $table->smallInteger('cmc')->nullable();
+            $table->mediumInteger('cmc')->nullable();
             $table->string('type_line')->nullable();
             $table->string('skryfall_image_small')->nullable();
             $table->string('skryfall_image_normal')->nullable();
